f5-vnf-ha-failover/Failover Script using Code Engine/requirements.txt
flask
ibm-cloud-sdk-core
ibm-vpc>=0.17.0
python-dotenv